refactor(rimap-config): extract validate/compose.rs from validate/mod.rs
Moves ValidatedAccountConfig, ValidatedMultiConfig,
ValidateAccountInputs, the four entry points (validate_multi,
validate_multi_allowing_empty, validate_legacy_as_multi,
validate_multi_inner), the validate_account orchestrator, and the
tests module from validate/mod.rs (970 LOC) to a new
validate/compose.rs sibling (958 LOC). validate/mod.rs shrinks to
32 LOC of pure wiring + re-exports.

mod.rs re-exports the four public symbols (Validated*Config,
validate_multi, validate_legacy_as_multi, and the feature-gated
validate_multi_allowing_empty) so external consumers
(crates/rimap-config/src/lib.rs and loader.rs) and the legacy
`crate::validate::*` paths in tests continue to resolve.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 3, hub 3/4)
